fix a problem with some unsigned int. The problem was found compiling on Windows by Ivan Dmitriev
git-svn-id: http://root.cern.ch/svn/root/trunk@13614 27541ba8-7e3a-0410-8455-c3a389f83636

+++ minuit2/src/mndaxpy.cxx
@@ -1,4 +1,4 @@
-// @(#)root/minuit2:$Name:  $:$Id: mndaxpy.cpp,v 1.1.6.3 2005/11/29 11:08:35 moneta Exp $
+// @(#)root/minuit2:$Name:  $:$Id: mndaxpy.cxx,v 1.1 2005/11/29 14:43:31 moneta Exp $
 // Authors: M. Winkler, F. James, L. Moneta, A. Zsenei   2003-2005  
 
 /**********************************************************************
@@ -53,10 +53,10 @@ int Mndaxpy(unsigned int n, double da, const double* dx, int incx, double* dy,
     ix = 1;
     iy = 1;
     if (incx < 0) {
-	ix = (-(n) + 1) * incx + 1;
+	ix = (-static_cast<int>(n) + 1) * incx + 1;
     }
     if (incy < 0) {
-	iy = (-(n) + 1) * incy + 1;
+	iy = (-static_cast<int>(n) + 1) * incy + 1;
     }
     i__1 = n;
     for (i__ = 1; i__ <= i__1; ++i__) {
